Rename setupTests to testUtils #346
