+++ src/api/units.test.tsx
@@ -1,6 +1,6 @@
 import { renderHook, waitFor } from '@testing-library/react';
 import UnitsJSON from '../mocks/units.json';
-import { hooksWrapperWithProviders } from '../setupTests';
+import { hooksWrapperWithProviders } from '../testUtils';
 import { useUnits } from './units';
 
 describe('useUnits', () => {
