+++ src/catalogue/category/catalogueCard.component.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderComponentWithBrowserRouter } from '../../setupTests';
+import { renderComponentWithBrowserRouter } from '../../testUtils';
 import { screen } from '@testing-library/react';
 import CatalogueCard, { CatalogueCardProps } from './catalogueCard.component';
 import userEvent from '@testing-library/user-event';
