+++ src/catalogue/category/catalogueCategoryDirectoryDialog.component.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderComponentWithBrowserRouter } from '../../setupTests';
+import { renderComponentWithBrowserRouter } from '../../testUtils';
 import { screen, waitFor } from '@testing-library/react';
 import CatalogueCategoryDirectoryDialog, {
   CatalogueCategoryDirectoryDialogProps,
