+++ src/catalogue/category/catalogueCategoryTableView.component.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderComponentWithBrowserRouter } from '../../setupTests';
+import { renderComponentWithBrowserRouter } from '../../testUtils';
 import { screen, waitFor } from '@testing-library/react';
 
 import userEvent from '@testing-library/user-event';
