+++ src/catalogue/items/catalogueItemsLandingPage.component.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderComponentWithMemoryRouter } from '../../setupTests';
+import { renderComponentWithMemoryRouter } from '../../testUtils';
 import { screen, waitFor } from '@testing-library/react';
 import CatalogueItemsLandingPage from './catalogueItemsLandingPage.component';
 import userEvent from '@testing-library/user-event';
