+++ src/catalogue/items/obsoleteCatalogueItemDialog.component.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {
   getCatalogueItemById,
   renderComponentWithBrowserRouter,
-} from '../../setupTests';
+} from '../../testUtils';
 
 import { fireEvent, screen, waitFor, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
