+++ src/items/itemsTable.component.test.tsx
@@ -3,7 +3,7 @@ import {
   getCatalogueCategoryById,
   getCatalogueItemById,
   renderComponentWithBrowserRouter,
-} from '../setupTests';
+} from '../testUtils';
 import userEvent from '@testing-library/user-event';
 import { waitFor, screen } from '@testing-library/react';
 import ItemsTable, { ItemTableProps } from './itemsTable.component';
