+++ src/view/breadcrumbs.component.test.tsx
@@ -2,7 +2,7 @@ import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 import Breadcrumbs, { BreadcrumbsProps } from './breadcrumbs.component';
-import { renderComponentWithBrowserRouter } from '../setupTests';
+import { renderComponentWithBrowserRouter } from '../testUtils';
 
 describe('Channel Breadcrumbs', () => {
   const onChangeNode = vi.fn();
