+++ src/view/viewTabs.component.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderComponentWithMemoryRouter } from '../setupTests';
+import { renderComponentWithMemoryRouter } from '../testUtils';
 import ViewTabs from './viewTabs.component';
 import { screen, waitFor, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
